+++ deps.ts
@@ -8,3 +8,13 @@ export type {
   Next,
   RouteContext,
 } from "jsr:@oak/oak@^16.1.0";
+
+export {
+  OpenApiGeneratorV3,
+  OpenAPIRegistry,
+  type ResponseConfig,
+  type RouteConfig as OasRouteConfig,
+  type ZodRequestBody,
+} from "npm:@asteasolutions/zod-to-openapi@^7.1.1";
+
+export { type OpenAPIObjectConfig } from "npm:@asteasolutions/zod-to-openapi@^7.1.1/dist/v3.0/openapi-generator";