+++ src/Delete.ts
@@ -1,12 +1,17 @@
 import { debug } from "./utils/logger.ts";
 import { register } from "./Store.ts";
 import { getUserSuppliedDecoratedMethodName } from "./utils/getUserSuppliedDecoratedMethodName.ts";
+import { type OasRouteConfig } from "../deps.ts";
+import { updateOas } from "./oasStore.ts";
 
 /**
  * Decorator that should be used on the Controller Class Method
  * for DELETE endpoints
  */
-export const Delete = (path: string = "") =>
+export const Delete = (
+  path: string = "",
+  openApiSpecs?: OasRouteConfig,
+) =>
 // deno-lint-ignore ban-types
 (arg1: Function | object, arg2: ClassMethodDecoratorContext | string): void => {
   const fnName: string = getUserSuppliedDecoratedMethodName(arg1, arg2);
@@ -16,6 +21,7 @@ export const Delete = (path: string = "") =>
     arg2,
   );
   register("delete", path, fnName);
+  updateOas(fnName, "delete", path, openApiSpecs);
 };
 
 export const _internal = { Delete };